% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{alpha_creator}
\alias{alpha_creator}
\title{alpha_creator}
\usage{
alpha_creator(
  y,
  kmat_dfs,
  alpha_prior_V,
  N,
  dist = "mvn",
  sigmasq = NULL,
  ald_tau = NULL,
  ald_theta = NULL,
  ald_z_vec = NULL,
  ald_z_mat = NULL,
  ald_bigB_inv = NULL
)
}
\arguments{
\item{kmat_dfs}{A list of dataframes of Gaussian kernel matrices for sets}

\item{N}{A scalar numeric for the number of iterations in the Gibbs sampler}

\item{distn}{A string for the choice of distribution "mvn" for multivariate normal (default). "ald" for asymmetric LaPlace.}
}
\value{
A list of matrices, each containing the initial kernel weights
}
\description{
Function to solve for initial kernel weights given kernel matrices and initialize a placeholder list of dataframes for kernel weights
}
