% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{bigB_creator}
\alias{bigB_creator}
\title{bigB_creator}
\usage{
bigB_creator(kmat_dfs, num_pw)
}
\arguments{
\item{kmat_dfs}{A list of dataframes of Gaussian kernel matrices for sets}

\item{num_pw}{A scalar numeric for the number of sets in the data}
}
\value{
A list of lists, each containing the diagonal entries of a kernel matrix
}
\description{
Function to extract diagonals from list of kernel matrices dataframes for asymmetric LaPlace B parameter
}
