% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McMCalgo.R
\name{BIGM}
\alias{BIGM}
\title{BIGM}
\usage{
BIGM(
  y,
  X,
  corrmat,
  num_locs,
  prior_scaler = 10,
  N = 1000,
  k_on = 0,
  rand = FALSE,
  lkli = "mvn",
  ald_skew = 0.5
)
}
\arguments{
\item{y}{A continuous response variable as a numeric vector.}

\item{X}{A dataframe or data matrix.}

\item{corrmat}{A matrix of prior interactions between elements. Can be set as the squared correlation.}

\item{num_locs}{A numeric value representing number of elements under examination.}

\item{prior_scaler}{A numeric value scaling the specified interaction matrix.}

\item{N}{A numeric value for number of McMC iterations.}

\item{k_on}{A logical value. FALSE (default). TRUE incorporates a matrix of marginal regressions as an external field effect -- see Fang & Kim 2016}

\item{rand}{A logical value. FALSE (default). TRUE randomizes the sampling order of the elements in the McMC exploration -- not recommended.}

\item{lkli}{A string. 'mvn' (default) uses a normal likelihood. 'ald' uses an asymmetric Laplace likelihood'}

\item{ald_skew}{A numeric value for the skew of the asymmetric LaPlace likelihood -- default 0.5, median.}
}
\value{
A binary vector containing selection status for elements within sets.
}
\description{
McMC version of the MLNR algorithm.
}
