% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{compute_correlation}
\alias{compute_correlation}
\title{compute_correlation}
\usage{
compute_correlation(df1, df2)
}
\arguments{
\item{df1}{A dataframe containing the first set}

\item{df2}{A dataframe containing the second set}
}
\value{
A scalar value for the average of correlations between the input dataframes
}
\description{
Function to compute correlation between two sets
}
