% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{corr_mat_creator}
\alias{corr_mat_creator}
\title{corr_mat_creator}
\usage{
corr_mat_creator(df_list, num_pw)
}
\arguments{
\item{df_list}{A list of dataframes of sets}

\item{num_pw}{A scalar numeric for the number of sets in the data}
}
\value{
A list of set dataframes
}
\description{
Function to create a list of correlation dataframes each holding correlations between elements of a set, one for each set.
}
