% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNetworkViz.R
\name{CreateNetworkViz}
\alias{CreateNetworkViz}
\title{Create Network Visualizations}
\usage{
CreateNetworkViz(
  dat,
  MLN_results,
  num_sets,
  set_names = c("Alanine and aspartate metabolism", "ATP synthesis",
    "c22_U133_probes(user defined)", "c23_U133_probes(user defined)",
    "JAK-Stat Signaling Pathway", "Oxidative phosphorylation",
    "OXPHOS_HG-U133A_probes(user defined)", "Parkinson's disease",
    "Ubiquinone biosynthesis", "gamma-Hexachlorocyclohexane degradation")
)
}
\arguments{
\item{dat}{Data frame containing the data to visualize.}

\item{MLN_results}{Results from the multilevel network recovery model.}

\item{num_sets}{Number of sets (pathways) to visualize.}

\item{set_names}{A character vector containing names of the sets (pathways) to be used in the visualization.}
}
\value{
A ggplot object with the network visualization.
}
\description{
This function creates a network visualization based on pathway data and machine learning results.
It generates a network of elements (genes, proteins, etc.) with edges representing interactions or relationships.
}
