% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{curr_xi_creator}
\alias{curr_xi_creator}
\title{curr_xi_creator}
\usage{
curr_xi_creator(pwy_dfs, N)
}
\arguments{
\item{pwy_dfs}{A list of dataframes of sets}

\item{N}{A scalar numeric for the number of iterations in the Gibbs sampler}
}
\value{
A list of matrices containing the initial inclusion variables for each set's elements
}
\description{
Function to initialize element inclusion variables and create placeholders for inclusion variables
}
