% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{dcorr}
\alias{dcorr}
\title{dcorr}
\usage{
dcorr(c2, n, ncp = 0)
}
\arguments{
\item{c2}{A numeric vector containing squared correlation values.}

\item{n}{the number of observations.}

\item{ncp}{(default=0) A logical value to specify the non-centrality parameter if desired. Seeting a non-zero `ncp` switches to an F-distribution while zero is a beta distribution}
}
\value{
a numeric value for density.
}
\description{
Function to compute density of the estimated squared-correlation statistic
}
