% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearDemoCode.R
\name{demo_code_Linear}
\alias{demo_code_Linear}
\title{Linear Model Demo Code}
\usage{
demo_code_Linear(
  num_sets = 5,
  n_samples = 100,
  distn_mean = 0,
  distn_sd = 1,
  skew = 0.5,
  sigmasq_y = 1,
  a = 1,
  b = 1,
  penalty = "function",
  dist = "mvn",
  mthd = "VB",
  Restarts = 1,
  rel_method = "gp",
  w_set = "avg.corr"
)
}
\arguments{
\item{num_sets}{Number of top-level sets.}

\item{n_samples}{Number of observations.}

\item{distn_mean}{Mean for the distribution (default 0).}

\item{distn_sd}{Standard deviation for the distribution (default 1).}

\item{skew}{Skewness for the distribution (default 0.5).}

\item{sigmasq_y}{Variance for the output variable (default 1).}

\item{a, }{b Hyperparameters for the model.}

\item{penalty}{Type of penalty for the model (default "function").}

\item{dist}{Distribution type (default "mvn").}

\item{mthd}{Method for fitting (default "VB").}

\item{Restarts}{Number of restarts for the fitting algorithm (default 1).}

\item{rel_method}{Method for relevance (default "gp").}

\item{w_set}{Weighting scheme (default "avg.corr").}
}
\value{
A list containing model results, intermediate outputs, and visualizations.
}
\description{
This function demonstrates how to generate simulated data, fit an MLNR model,
and visualize the results.
}
