% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSimulationData.R
\name{GenSimDat}
\alias{GenSimDat}
\title{Generate Simulated Data}
\usage{
GenSimDat(
  distn,
  num_samples,
  num_sets,
  distn_mean = 0,
  distn_sd = 1,
  skew = 0.5
)
}
\arguments{
\item{distn}{Character. Distribution type ("mvn" for Gaussian noise, "ald" for asymmetric Laplace noise).}

\item{num_samples}{Integer. Number of samples to generate.}

\item{num_sets}{Integer. Number of pathways to generate.}

\item{distn_mean}{Numeric. Mean of the distribution. Default is 0.}

\item{distn_sd}{Numeric. Standard deviation of the distribution. Default is 1.}

\item{skew}{Numeric. Skewness parameter for the asymmetric Laplace distribution. Default is 0.5.}
}
\value{
A data frame containing the generated data.
}
\description{
This function generates simulated data based on specified distribution types
and a customizable number of features and paths.
}
\details{
The function generates features for multiple pathways using multivariate normal distributions.
The first two pathways have predefined covariance structures, while additional pathways
are generated dynamically with random correlation matrices.
}
\examples{
# Generate data with Gaussian noise
data <- GenSimDat("mvn", num_samples = 100, num_sets = 3)

# Generate data with asymmetric Laplace noise
data <- GenSimDat("ald", num_samples = 100, num_sets = 3, skew = 0.7)

}
