% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{kmat_creator}
\alias{kmat_creator}
\title{kmat_creator}
\usage{
kmat_creator(df_list, num_pw)
}
\arguments{
\item{df_list}{A list of dataframes of sets}

\item{num_pw}{A scalar numeric for the number of sets in the data}
}
\value{
A list of set dataframes
}
\description{
Function to create a list of kernel matrices dataframes each holding the Gaussian kernel matrix from covariates of a set, one for each set.
}
