% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{log_ratio_d_e}
\alias{log_ratio_d_e}
\title{log_ratio_d_e}
\usage{
log_ratio_d_e(log_d, log_e)
}
\arguments{
\item{log_d}{Numeric vector..}

\item{log_e}{Numeric vector.}
}
\value{
A numeric value.
}
\description{
Computes the log-sum-exp trick for two vectors simultaneously.
}
