% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{log_sum_exp}
\usage{
log_sum_exp(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric value.
}
\description{
Computes the log-sum-exponential trrick.
}
