% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{min_max_normalization_df}
\alias{min_max_normalization_df}
\title{min_max_normalization_df}
\usage{
min_max_normalization_df(df)
}
\arguments{
\item{df}{A dataframe containing numeric values}
}
\description{
Function to apply min-max normalization to all columns of a dataframe
}
