% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLNRpredict.R
\name{MLNR.predict}
\alias{MLNR.predict}
\title{Prediction Wrapper for MLNR}
\usage{
MLNR.predict(dat_pred, model, cov_transform = "scale")
}
\arguments{
\item{dat_pred}{Data frame containing the covariates for which the response is to be predicted. Should be formatted similarly to the input data to the model, with the first column dropped.}

\item{model}{List output of function `MLNR` containing all the saved model parameters.}

\item{cov_transform}{Character string specifying the transformation to apply to covariates. Default is "scale".}
}
\value{
A vector containing predicted values of the response.
}
\description{
This function predicts a continuous response using the MLNR model for the input dataset.
}
