% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R, R/MLNRfunction.R
\name{MLNR}
\alias{MLNR}
\title{MLNR}
\usage{
MLNR(
  dat,
  num_pwy,
  skipper = 300,
  smpl.sz = 2,
  N_norm = 2000,
  level_1_connected = 1,
  sigmasq_y = 1,
  a = 1,
  b = 1,
  ald_p = 0.5,
  n0 = 1,
  s0 = 1,
  pi_0 = 0.5,
  a_al = 0.5,
  b_al = 0.5,
  sigmasq_alpha = 1,
  penalty = "weights",
  dist = "mvn",
  mthd = "VB",
  Restarts = 1,
  rel_method = "mi",
  w_set = "avg.corr"
)
}
\arguments{
\item{dat}{Input dataset.}

\item{num_pwy}{Number of pathways.}

\item{skipper}{Number of Gibbs sampling iterations for the kernel weights between each element selection run.}

\item{smpl.sz}{Sample size of pathways tested at once.}

\item{N_norm}{Number of total iterations for normalization.}

\item{level_1_connected}{If 1, supernodes are connected; if 0, they are not.}

\item{sigmasq_y}{Initial overall variance.}

\item{a}{Shape parameter for the inverse gamma prior on the overall variance.}

\item{b}{Scale parameter for the inverse gamma prior on the overall variance.}

\item{ald_p}{Asymmetric Laplace Distribution (ALD) quantile parameter.}

\item{n0}{ALD scale prior parameter.}

\item{s0}{ALD scale prior parameter.}

\item{pi_0}{Binomial probability for pathway selection.}

\item{a_al}{Shape parameter for the inverse gamma prior on the coefficient variance.}

\item{b_al}{Scale parameter for the inverse gamma prior on the coefficient variance.}

\item{sigmasq_alpha}{Initial variance for coefficients.}

\item{penalty}{Type of penalty, either "weights" (penalizes only kernel weights) or "function" (penalizes function complexity).}

\item{dist}{Distribution type, either "mvn" (multivariate normal) or "ald" (asymmetric Laplace).}

\item{mthd}{Method for inference, either "VB" (variational Bayes) or "MCMC" (Markov Chain Monte Carlo).}

\item{Restarts}{Number of restarts to help the Ising model escape local modes.}

\item{rel_method}{Relevance method, either "gp" (uses actual observations), "mi" (mutual information), or "cor" (correlation squared).}

\item{w_set}{Weight setting, such as "avg.cor" for using average correlation.}

\item{reg_results}{If 1, the model samples final weights given selected elements and sets, and computes MSE.}
}
\value{
MLNR model results.
}
\description{
Helper functions for Bayesian Multilevel Network Recovery

Implements the MLNR model for Bayesian inference in multilevel network regression.
}
