% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{nlsep}
\alias{nlsep}
\title{nlsep}
\usage{
nlsep(par, X, Y)
}
\arguments{
\item{par}{Parameter vector for constructed GP.}

\item{X}{A dataframe or data matrix.}

\item{Y}{A continuous response variable as a numeric vector.}
}
\value{
A numeric value for negative log-likelihood.
}
\description{
Computes seperable negative log-likelihood --see Gramacy 2020
}
