% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{pathway_creator}
\alias{pathway_creator}
\title{pathway_creator}
\usage{
pathway_creator(dat, num_pw, transform = "scale")
}
\arguments{
\item{dat}{A dataframe containing the data in its perferred format. Response value in first column, set number in last row.}

\item{num_pw}{A scalar numeric for the number of sets in the data}

\item{transform}{(default="scale") A string value indicating the desired tranformation on the covariates. "sclae" for a scale transformation. "minmax" for min-max normalization}
}
\value{
A list of set dataframes
}
\description{
Function to create a list of dataframes, one for each set.
}
