% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{pathway_creator_pred}
\alias{pathway_creator_pred}
\title{pathway_creator_pred}
\usage{
pathway_creator_pred(dat_pred, dat, num_pw, transform)
}
\arguments{
\item{dat}{A dataframe containing the data in its perferred format. Response value in first column, set number in last row.}

\item{num_pw}{A scalar numeric for the number of sets in the data}

\item{transform}{(default="scale") A string value indicating the desired tranformation on the covariates. "sclae" for a scale transformation. "minmax" for min-max normalization}

\item{mu}{mean of training covariates}

\item{sigma}{standard deviation of training covariates}

\item{xmin}{minimum of training covariates}

\item{xmax}{maximum of training covariates}
}
\value{
A list of set dataframes
}
\description{
Function to create a list of dataframes, one for each set, using the scaling of the training data
}
