% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{prob_notnull_creator}
\alias{prob_notnull_creator}
\title{prob_notnull_creator}
\usage{
prob_notnull_creator(pwy_dfs, N, init = 1)
}
\arguments{
\item{pwy_dfs}{A list of dataframes of sets}

\item{N}{A scalar numeric for the number of iterations in the Gibbs sampler}

\item{init}{A numeric value for the initial probability of inclusion of sets. Default is 1}
}
\value{
A list of lists containing the initial inclusion probability for each set
}
\description{
Function to initialize set probabilities and create placeholders for computed probabilities
}
