% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{prv.alpha}
\alias{prv.alpha}
\title{prv.alpha}
\usage{
prv.alpha(x, i)
}
\arguments{
\item{x}{A matrix containing sampled numeric values}
}
\value{
A numeric vector of last sampled values
}
\description{
Function to extract last sampled value from a matrix of sampled values. Finds the last sampled non-NA entry for each column
}
