% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{sigmoid}
\alias{sigmoid}
\title{sigmoid}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{A numeric vector containing the values to which we want to apply a sigmoid transformation.}
}
\value{
a transformed version of the input.
}
\description{
Function to compute the sigmoid
}
