% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBalgo.R
\name{VB}
\alias{VB}
\title{VB}
\usage{
VB(
  y,
  X,
  corrmat,
  num_locs,
  lambda = 1,
  iters = 50,
  GP = 0,
  Sigmat = NULL,
  prior_scaler = 10,
  plot_on = FALSE,
  true_dist = 0,
  ker_wts = NULL,
  rand = TRUE,
  lambda_cv = 2,
  lkli = "mvn"
)
}
\arguments{
\item{y}{A continuous response variable as a numeric vector.}

\item{X}{A dataframe or data matrix.}

\item{corrmat}{A matrix of prior interactions between elements. Can be set as the squared correlation.}

\item{num_locs}{A numeric value representing number of elements under examination.}

\item{iters}{A numeric value for number of VB iterations.}

\item{prior_scaler}{A numeric value scaling the specified interaction matrix.}

\item{rand}{A logical value. FALSE (default). TRUE randomizes the sampling order of the elements in the McMC exploration -- not recommended.}

\item{lkli}{A string. 'mvn' (default) uses a normal likelihood. 'ald' uses an asymmetric Laplace likelihood'}

\item{k_on}{A logical value. FALSE (default). TRUE incorporates a matrix of marginal regressions as an external field effect -- see Fang & Kim 2016}

\item{ald_skew}{A numeric value for the skew of the asymmetric LaPlace likelihood -- default 0.5, median.}
}
\value{
A binary vector containing selection status for elements within sets.
}
\description{
Variational Bayes version of the MLNR algorithm.
}
