% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNetworkViz.R
\name{CreateNetworkViz}
\alias{CreateNetworkViz}
\title{Create Network Visualizations}
\usage{
CreateNetworkViz(dat, gamma, xi, num_sets, set_names = NULL)
}
\arguments{
\item{dat}{Data frame containing the data to visualize.}

\item{gamma}{set Results from the multilevel network recovery model.}

\item{xi}{element Results from the multilevel network recovery model.}

\item{num_sets}{Number of sets (pathways) to visualize.}

\item{set_names}{A character vector containing names of the sets (pathways) to be used in the visualization.}
}
\value{
A ggplot object with the network visualization.
}
\description{
This function creates a network visualization based on set data and machine learning results.
It generates a network of elements (genes, proteins, etc.) with edges representing interactions or relationships.
}
